+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/model/CPOptionValueSoap.java
@@ -187,11 +187,11 @@ public class CPOptionValueSoap implements Serializable {
 		_title = title;
 	}
 
-	public int getPriority() {
+	public double getPriority() {
 		return _priority;
 	}
 
-	public void setPriority(int priority) {
+	public void setPriority(double priority) {
 		_priority = priority;
 	}
 
@@ -214,6 +214,6 @@ public class CPOptionValueSoap implements Serializable {
 	private long _CPOptionId;
 	private String _name;
 	private String _title;
-	private int _priority;
+	private double _priority;
 	private Date _lastPublishDate;
 }
